+++ experiment.kv
@@ -8,33 +8,29 @@ ScreenManager:
     id: screen_manager
     transition: sm.NoTransition()
 
-#    WelcomeScreen:
-#        id: welcome_screen
-#        name: "WelcomeScreen"
-#        manager: screen_manager
-#
-#    DataScreen:
-#        name: "DataScreen"
-#        manager: screen_manager
-#        label_text: welcome_screen.text
-#
-#    StartScreen:
-#        name: "StartScreen"
-#        manager: screen_manager
-#
-#    AgreementScreen:
-#        id: agreement_screen
-#        name: "AgreementScreen"
-#        manager: screen_manager
-#
-#    SocioLingScreen:
-#        name: "SocioLingScreen"
-#        manager: screen_manager
-#        email: agreement_screen.email
-#
-#    AttentionScreen:
-#        name: "AttentionScreen"
-#        manager: screen_manager
+    WelcomeScreen:
+        id: welcome_screen
+        name: "WelcomeScreen"
+        manager: screen_manager
+
+    DataScreen:
+        name: "DataScreen"
+        manager: screen_manager
+        label_text: welcome_screen.text
+
+    StartScreen:
+        name: "StartScreen"
+        manager: screen_manager
+
+    AgreementScreen:
+        id: agreement_screen
+        name: "AgreementScreen"
+        manager: screen_manager
+
+    SocioLingScreen:
+        name: "SocioLingScreen"
+        manager: screen_manager
+        email: agreement_screen.email
 
     TrainingScreen:
         name: "TrainingScreen"
@@ -106,23 +102,6 @@ ScreenManager:
             text: "Start"
             font_size: 50
             on_press: root.manager.current = "AgreementScreen"
-            background_color: (0, 1, 0, 1)
-
-<AttentionScreen>:
-    BoxLayout:
-        orientation: "vertical"
-
-        Label:
-            color: 0, 0, 0, 1
-            text: "You’re about to enter the experiment. \nPlease turn off your cell phone :)"
-            font_size: 40
-
-        Button:
-            size_hint_y: None
-            height: 100
-            text: "Continue"
-            font_size: 50
-            on_press: root.manager.current = "TrainingScreen"
             background_color: (0, 1, 0, 1)
 
 <LoadDialog>:
