+++ experiment_boxes/acceptability_box.kv
@@ -1,0 +1,58 @@
+<AcceptabilityBox>:
+    orientation: "vertical"
+    canvas.before:
+        Color:
+            rgba: (1, 1, 1, 1)
+        Rectangle:
+            pos: self.pos
+            size: self.size
+
+    ColoredLabel:
+        text: "Please rate the correctness of the following sentence on a scale from 1 (ungrammatical) to 5 (perfectly grammatical)"
+        text_size: self.width, None
+        font_size: 25
+        size_hint: 1, None
+        height: self.texture_size[1] + 15
+        halign: "center"
+
+    RelativeLayout:
+        id: progress_bar
+        size_hint: (None, None)
+        height: 75
+        width: root.width * 0.2
+        pos_hint: {'center_x': 1}
+
+    GridLayout:
+        cols: 5
+        row_force_default: True
+        row_default_height: 40
+
+        RatingButton:
+            text: "1"
+            on_press: root.experiment_design.record_active_state(self.state, self.text)
+
+        RatingButton:
+            text: "2"
+            on_press: root.experiment_design.record_active_state(self.state, self.text)
+
+        RatingButton:
+            text: "3"
+            on_press: root.experiment_design.record_active_state(self.state, self.text)
+
+        RatingButton:
+            text: "4"
+            on_press: root.experiment_design.record_active_state(self.state, self.text)
+
+        RatingButton:
+            text: "5"
+            on_press: root.experiment_design.record_active_state(self.state, self.text)
+
+    Widget:
+
+    Button:
+        size_hint_y: None
+        text: "Continue"
+        height: 40
+        font_size: 25
+        background_color: (0.608, 0.608, 0.627, 0.5)
+        on_press: root.experiment_design.record_result()
