+++ experiment_screen.py
@@ -1,0 +1,8 @@
+from kivy.uix.screenmanager import Screen
+
+
+class ExperimentScreen(Screen):
+    """
+    Where actual experiment takes place
+    """
+    pass
