+++ experiment_types/self_paced_design.py
@@ -1,0 +1,11 @@
+
+class SelfPacedDesign:
+    instructions = "You will see a series of dashes representing the letters in each word, " \
+                   "with spaces showing the gaps between words. \n\n" \
+                   "Press [b]ENTER[/b] or [b]SPACEBAR[/b] to see the first word in the sentence. \n" \
+                   "Then press the key once more, and the next word will be displayed. \n" \
+                   "That is, you will only see one word displayed at a time. \n" \
+                   "After each sentence you will be asked a simple question about the sentence. \n" \
+                   "Your task is to read the sentence as quickly as possible, " \
+                   "while understanding it, as checked by the question."
+
