+++ training.kv
@@ -1,61 +1,4 @@
 <TrainingScreen>:
-    id: training
-
-    BoxLayout:
-        id: sentence_box
-        orientation: "vertical"
-
-        ColoredLabel:
-            text: "Please rate the correctness of the following sentence on a scale from 1 (ungrammatical) to 5 (perfectly grammatical)"
-            text_size: self.width, None
-            font_size: 25
-            size_hint: 1, None
-            height: self.texture_size[1] + 15
-            halign: "center"
-
-        RelativeLayout:
-            id: progress_bar
-            size_hint: (None, None)
-            height: 75
-            width: root.width * 0.2
-            pos_hint: {'center_x': 1}
-
-        GridLayout:
-            cols: 5
-            row_force_default: True
-            row_default_height: 40
-
-            RatingButton:
-                text: "1"
-                on_press: root.record_active_state(self.state, self.text)
-
-            RatingButton:
-                text: "2"
-                on_press: root.record_active_state(self.state, self.text)
-
-            RatingButton:
-                text: "3"
-                on_press: root.record_active_state(self.state, self.text)
-
-            RatingButton:
-                text: "4"
-                on_press: root.record_active_state(self.state, self.text)
-
-            RatingButton:
-                text: "5"
-                on_press: root.record_active_state(self.state, self.text)
-
-        Widget:
-
-        Button:
-            size_hint_y: None
-            text: "Continue"
-            height: 40
-            font_size: 25
-            background_color: (0.608, 0.608, 0.627, 0.5)
-            on_press: root.record_result()
-
-
     BoxLayout:
         id: instruction_box
         orientation: "vertical"
@@ -67,9 +10,10 @@
                 size: self.size
         Widget:
         Label:
+            id: instructions_label
             color: 0, 0, 0, 1
-            text: training.instructions
             font_size: 30
+            markup: True
             text_size: self.width, None
             size_hint: 1, None
             height: self.texture_size[1]
@@ -83,6 +27,33 @@
             font_size: 50
             background_color: (0, 1, 0, 1)
             on_press:
-                training.remove_widgets(["instruction_box"])
-                training.display_sentence()
+                root.remove_widgets(["instruction_box"])
+                root.experiment_design.display_sentence()
+
+
+    BoxLayout:
+        id: attention_box
+        orientation: "vertical"
+        canvas.before:
+            Color:
+                rgba: (1, 1, 1, 1)
+            Rectangle:
+                pos: self.pos
+                size: self.size
+
+        Label:
+            color: 0, 0, 0, 1
+            text: "You’re about to enter the experiment. \nPlease turn off your cell phone :)"
+            font_size: 40
+
+        Button:
+            size_hint_y: None
+            height: 100
+            text: "Continue"
+            font_size: 50
+            background_color: (0, 1, 0, 1)
+            on_press:
+                root.remove_widgets(["attention_box"])
+                root.load_experiment()
+
